+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v0.1.13] - 2018-10-26
+### Added
+- Add dc.theme-dev.yml file to expose containers theme files
+- Add run.sh script to start and stop Docker containers
+
 ## [v0.1.12] - 2018-10-23
 ### Changed
 - Fix `.gitignore` pattern